extract prepare helpers into eagle_worker_common
